[#32] Multiple alert times for all entries
As for now, org-wild-notifier notifies just once, if
WILD_NOTIFIER_NOTIFY_BEFORE property is not set. If user wants to be
notified of all events multiple times, they'd have to set the property
on the every single agenda entry, which is inconvenient.

This change tweaks `org-wild-notifier-alert-time` custom variable to
support array values, which effectively gives the semantic described
above.