Improve notification timing.
Notification is only sent when the current time is within the minute of the set interval
from the event time. This usually works if the check runs reliably every 60 seconds.

However, Emacs doesn't guarantee that the timer fires exactly every time. Blocking
operations can delay the timer significantly, sometimes up to tens of seconds late.

This commit changes the notification timing slightly to favor notifying just *before* the
set interval so it's not too late. For example, if the current time is 12 seconds before the
set interval from the event time, it will fire the notification instead of waiting for the
next check which will be 48 seconds late.

In order to not duplicate notifications, a new variable org-wild-notifier--last-check-time
is added to remember the last time the check ran. If a nofication was fired in the last
check, it won't fire it again.

With this behavior change, it can even support alert time down to 0 minute from the event
time.