Fix bug in check event.
I introduced a bug in 68c34ae when changing the way event is checked for notification. The
last check time was updated on the first event check instead of at the end of all event
checks. That meant it would only properly fire notification for the first event. This commit
fixes the bug by moving the last check time update to the end of the all event checks.