Run async process in user-emacs-directory (#68)
By default, `async-start' will use the current buffer's
default-directory as the working directory for the spawned Emacs
process.  In a timer, though, the current buffer may be whatever the
user happens to be looking at, and the directory may not exist (which
causes annoying error messages).

Hard code the process to run in `user-emacs-directory', which ought to
always exist and should be portable across operating systems.